Expose/hide types that need to be exposed/hidden
