+++ lib/blemulator/internal/simulation_manager_mixins/response_models.dart
@@ -1,0 +1,8 @@
+part of internal;
+
+class CharacteristicResponse {
+  SimulatedCharacteristic characteristic;
+  Uint8List value;
+
+  CharacteristicResponse(this.characteristic, this.value);
+}
