+++ lib/blemulator/simulation/simulated_characteristic.dart
@@ -56,10 +56,3 @@ class SimulatedCharacteristic {
     return _streamController.stream;
   }
 }
-
-class CharacteristicResponse {
-  SimulatedCharacteristic characteristic;
-  Uint8List value;
-
-  CharacteristicResponse(this.characteristic, this.value);
-}
