+++ lib/internal/constants.dart
@@ -97,19 +97,6 @@ abstract class ArgumentName {
   static const String deviceIdentifiers = "deviceIdentifiers";
 }
 
-abstract class ScanMode {
-  static const int opportunistic = -1;
-  static const int lowPower = 0;
-  static const int balanced = 1;
-  static const int lowLatency = 2;
-}
-
-abstract class CallbackType {
-  static const int allMatches = 1;
-  static const int firstMatch = 2;
-  static const int matchLost = 4;
-}
-
 abstract class NativeConnectionState {
   static const String connecting = "connecting";
   static const String connected = "connected";
@@ -117,11 +104,4 @@ abstract class NativeConnectionState {
   static const String disconnected = "disconnected";
 }
 
-enum BluetoothState {
-  UNKNOWN,
-  UNSUPPORTED,
-  UNAUTHORIZED,
-  POWERED_ON,
-  POWERED_OFF,
-  RESETTING,
-}
+
